+++ src/kubeseal_auto/host.py
@@ -52,7 +52,7 @@ class Host:
         with requests.get(url) as r:
             if r.status_code == 404:
                 click.echo(f"The required version {version} is not available")
-                exit(1)
+                raise FileNotFoundError
             with open(local_path, "wb") as f:
                 f.write(r.content)