+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fallback to the default kubeseal binary if the specific version can't be downloaded
 ### Fixed
 - Correct exit code if the required version of kubeseal was not found on github